+++ source/Private/Get-ClassAst.ps1
@@ -2,6 +2,9 @@
     .SYNOPSIS
         Returns the AST for a single or all classes.
 
+    .DESCRIPTION
+        Returns the AST for a single or all classes.
+
     .PARAMETER ScriptFile
         The path to the source file that contain the class.
 
@@ -21,6 +24,7 @@
 function Get-ClassAst
 {
     [CmdletBinding()]
+    [OutputType([System.Collections.Generic.IEnumerable`1[System.Management.Automation.Language.Ast]])]
     param
     (
         [Parameter(Mandatory = $true)]