+++ source/Public/New-DscResourceWikiPage.ps1
@@ -21,6 +21,9 @@
         The path to the root of the built DSC resource module, e.g.
         'output/MyResource/1.0.0'.
 
+    .PARAMETER Force
+        Overwrites any existing file when outputting the generated content.
+
     .EXAMPLE
         New-DscResourceWikiPage `
             -SourcePath C:\repos\MyResource\source `
@@ -31,6 +34,7 @@
 #>
 function New-DscResourceWikiPage
 {
+    [Diagnostics.CodeAnalysis.SuppressMessageAttribute('PSUseShouldProcessForStateChangingFunctions', '')]
     [CmdletBinding()]
     param
     (