+++ CHANGELOG.md
@@ -25,6 +25,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
     `Required` or `Write`.
   - `Get-CompositeResourceParameterValidateSet` - Returns the array of values
     contained in the ValidateSet parameter attributes if it exists.
+- Added QA test to do some quality checks on the module code and change log.
 
 ### Changed
 
@@ -43,6 +44,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
     use with composite DSC resources.
   - Enabled the function to extract the comment block if it is not at the top
     of the script file to support composite resources.
+- Updated code to pass newly added quality checks.
 
 ### Fixed